photobooth: Add CameraOverlayPiCamera
Add a new type of camera 'overlaypicamera' that uses picamera
overlay to display preview.

This is a bit hackish but far smoother than the implementation in
CameraPicamera.py that takes 1/3 of a second to generate a picture
that is then rendered by PyQT resuting in a maximum of 3fps being
displayed in a preview.