Add a gui frame to setup RGB lightning color
This patch adds a gui frame to setup RGB lightning color.

This frame is able to drive a RGB ip lightning hardware (as of
today only yeelight hardware is supported).

It is possible to configure 10 colors trough the color_name_N
and color_rgb_N parameters in the PoseSetup configuration file.